\name{buildLibDB}
\alias{buildLibDB}
\title{Build a database of packages in a dev library}
\description{
Given a path to developer's library construct a data frame
with all the available packages with location and version information.
}
\usage{
buildLibDB(path = getOption("devlib"))
}
\arguments{
  \item{path}{character, path to the top level of the developer's library}
}
\value{
Data frame with columns
TODO
}
