\name{install.devlib}
\alias{install.devlib}
\title{
Install package to developer's library
}
\description{
Install a package to developer's library
}
\usage{
install.devlib(fname, pkg, ver, ...)
}
\arguments{
  \item{fname}{character, path to package file, tarball in *nix or zipped
  binary version on Windows}
  \item{pkg}{character, name of the package}
  \item{ver}{character, version of the package}
  \item{\dots}{other arguments passed to \code{\link{install.packages}}}
}
